Optional Boost for Open Transactions (Unix)
This commit makes Boost optional for UNIX.  Many bug-fixes are included in
this commit.
By default OT will compile with C++11 support, however it will by-default
use Boost.

If you wish to compile OT without Boost, you must use the latest version
of the in-development ChaiScript.  (Version 5).

In the future, the standard options will be changed so that Boost isn't
used by default.